Add integration-test for proxy/bridge
I've added a work-in-progress test case that launches an embedded Jetty
container with the ProxyServlet and event listeners installed.

A Felix container is started from one event listener, and the bridge-
and whiteboard bundles are installed (together with a bunch of
dependencies).

The minimal test framework needs to be refactored to allow more tests to
be written, but right now there's one test that fails becase of the
FELIX-5219 bug.