Add header that identifies proxied request
Since the Dispatcher handles both "normal" requests for the embedded
jetty container *and* for the proxy-servlet one cannot simply stop using
the `servletPath` when finding the correct servlet.

I've added a "hack" (needs to be refined) that uses a header to
distinguish requests that comes through the ProxyServlet. Those requests
should *never* use the `servletPath` when resolving a matching servlet.

Also, some other minor fixes:

- skip RAT checks for now
- add runtime dep on http.bridge to get a proper reactor order
- use the correct classifier for the jetty-all dep