Recalculate "servletPath" when dispatching a proxied request
For some libraries to work correctly when running as servlets inside an OSGi
container the HttpServletRequest must be consistent. This applied to
"osgi-jax-rs-connector" and to the "Felix WebConsole".

Basically, it must be possible to construct the correct request path by
concatening "contextPath", "servletPath" and "pathInfo" (so that it matches a
proper subset of the original requestURI).

In order to achieve this in the case of a proxied request, the servletPath that
should be propagated to servlets registered in the OSGi container is the
servletPath of the original request *+* the servletPath from the
"PathResolution".

Also note that this changeset includes a workaround for an issue introduced in
FELIX-5199.